fix product segment
